UI/Qt: Move dark mode check function to Qt GUI utils file
