+++ Ladybird/Qt/QtUtils.h
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2023, Tim Flynn <trflynn89@serenityos.org>
+ * Copyright (c) 2024, circl <circl.lastname@gmail.com>
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#pragma once
+
+#include <QWidget>
+
+namespace Ladybird {
+
+bool is_using_dark_system_theme(QWidget&);
+
+}
